final version after corrections
