+++ src/components/Pages/adminPanel/RequestsAdmin.tsx
@@ -2,9 +2,6 @@ import { useState } from 'react';
 import { useRequestStore } from '../../../store/RequestsStore';
 
 import {
-  Stack,
-  Modal,
-  Fade,
   Button,
   Box,
   Typography,
@@ -19,33 +16,21 @@ import {
 } from '@mui/material';
 import DeleteIcon from '@mui/icons-material/Delete';
 
-const style = {
-  position: 'absolute',
-  top: '50%',
-  left: '50%',
-  transform: 'translate(-50%, -50%)',
-  width: 400,
-  bgcolor: 'background.paper',
-  border: '2px solid #000',
-  boxShadow: 24,
-  p: 4,
-};
+import AdminConfirmModal from '../modals/AdminConfirmModal';
 
 export default function ReviewsAdmin() {
   const { requests, removeRequest } = useRequestStore();
 
   const [openConfirm, setOpenConfirm] = useState(false);
-  const [openStatus, setOpenStatus] = useState(false);
   const [itemToDelete, setItemToDelete] = useState<{
     id: string;
     type: 'request';
   } | null>(null);
-  const [statusMessage, setStatusMessage] = useState('');
 
   const handleRemoveReview = (id: string) => {
     removeRequest(id);
-    setStatusMessage('Заявка успешно удалена');
-    setOpenStatus(true);
+    setOpenConfirm(false);
+    setItemToDelete(null);
   };
 
   const handleOpenConfirm = (item: { id: string; type: 'request' }) => {
@@ -56,11 +41,6 @@ export default function ReviewsAdmin() {
   const handleCloseConfirm = () => {
     setOpenConfirm(false);
     setItemToDelete(null);
-  };
-
-  const handleCloseStatus = () => {
-    setOpenStatus(false);
-    setStatusMessage('');
   };
 
   const handleConfirmDelete = () => {
@@ -119,101 +99,39 @@ export default function ReviewsAdmin() {
         </TableContainer>
       </Box>
 
-      <Modal
+      <AdminConfirmModal
         open={openConfirm}
         onClose={handleCloseConfirm}
-        closeAfterTransition
+        title="Подтверждение удаления"
+        description="Вы уверены, что хотите удалить этот элемент?"
       >
-        <Fade in={openConfirm}>
-          <Box sx={style}>
-            <Typography variant="h6" component="h2">
-              Подтверждение удаления
-            </Typography>
-            <Typography sx={{ mt: 2 }}>
-              Вы уверены, что хотите удалить этот элемент?
-            </Typography>
-            <Stack
-              sx={{
-                flexDirection: 'row',
-                justifyContent: 'center',
-                gap: 2,
-                mt: 4,
-              }}
-            >
-              <Button
-                onClick={handleCloseConfirm}
-                sx={{
-                  backgroundColor: '#090E16',
-                  color: 'white',
-                  borderRadius: '16px',
-                  '&:hover': {
-                    backgroundColor: '#7B4AE2',
-                    '& .MuiLink-root': {
-                      color: '#7B4AE2',
-                      '& .MuiTypography-root': {
-                        color: '#7B4AE2',
-                      },
-                    },
-                  },
-                }}
-              >
-                Отмена
-              </Button>
-              <Button
-                onClick={handleConfirmDelete}
-                sx={{
-                  backgroundColor: '#090E16',
-                  color: 'white',
-                  borderRadius: '16px',
-                  '&:hover': {
-                    backgroundColor: '#7B4AE2',
-                    '& .MuiLink-root': {
-                      color: '#7B4AE2',
-                      '& .MuiTypography-root': {
-                        color: '#7B4AE2',
-                      },
-                    },
-                  },
-                }}
-              >
-                Удалить
-              </Button>
-            </Stack>
-          </Box>
-        </Fade>
-      </Modal>
-
-      <Modal open={openStatus} onClose={handleCloseStatus} closeAfterTransition>
-        <Fade in={openStatus}>
-          <Box sx={style}>
-            <Typography variant="h6" component="h2">
-              Статус удаления
-            </Typography>
-            <Typography sx={{ mt: 2 }}>{statusMessage}</Typography>
-            <Box sx={{ display: 'flex', justifyContent: 'center', mt: 4 }}>
-              <Button
-                onClick={handleCloseStatus}
-                sx={{
-                  backgroundColor: '#090E16',
-                  color: 'white',
-                  borderRadius: '16px',
-                  '&:hover': {
-                    backgroundColor: '#7B4AE2',
-                    '& .MuiLink-root': {
-                      color: '#7B4AE2',
-                      '& .MuiTypography-root': {
-                        color: '#7B4AE2',
-                      },
-                    },
-                  },
-                }}
-              >
-                Закрыть
-              </Button>
-            </Box>
-          </Box>
-        </Fade>
-      </Modal>
+        <Button
+          onClick={handleCloseConfirm}
+          sx={{
+            backgroundColor: '#090E16',
+            color: 'white',
+            borderRadius: '16px',
+            '&:hover': {
+              backgroundColor: '#7B4AE2',
+            },
+          }}
+        >
+          Отмена
+        </Button>
+        <Button
+          onClick={handleConfirmDelete}
+          sx={{
+            backgroundColor: '#090E16',
+            color: 'white',
+            borderRadius: '16px',
+            '&:hover': {
+              backgroundColor: '#7B4AE2',
+            },
+          }}
+        >
+          Удалить
+        </Button>
+      </AdminConfirmModal>
     </>
   );
 }
