+++ src/components/Pages/adminPanel/ReviewsAdmin.tsx
@@ -8,47 +8,11 @@ import {
   Typography,
   TextField,
   IconButton,
-  Modal,
-  Fade,
 } from '@mui/material';
-import { styled } from '@mui/material/styles';
+
 import DeleteIcon from '@mui/icons-material/Delete';
 
-const StyledTextField = styled(TextField)(() => ({
-  '& .MuiInputBase-root': {
-    color: '#ffffff',
-  },
-  '& .MuiOutlinedInput-notchedOutline': {
-    borderColor: '#ffffff80',
-  },
-  '&:hover .MuiOutlinedInput-notchedOutline': {
-    borderColor: '#ffffff80',
-  },
-  '&.Mui-focused .MuiOutlinedInput-notchedOutline': {
-    borderColor: '#ffffff80',
-  },
-  '& .MuiInputLabel-root': {
-    color: '#ffffff80',
-  },
-  '& .MuiInputLabel-root.Mui-focused': {
-    color: '#ffffff80',
-  },
-  '& .MuiInputLabel-root.MuiInputLabel-shrink': {
-    color: '#ffffff80',
-  },
-}));
-
-const style = {
-  position: 'absolute',
-  top: '50%',
-  left: '50%',
-  transform: 'translate(-50%, -50%)',
-  width: 400,
-  bgcolor: 'background.paper',
-  border: '2px solid #000',
-  boxShadow: 24,
-  p: 4,
-};
+import ConfirmModal from '../modals/AdminConfirmModal';
 
 export default function ReviewsAdmin() {
   const { reviews, addReview, removeReview } = useReviewsStore();
@@ -62,26 +26,15 @@ export default function ReviewsAdmin() {
   });
 
   const [openConfirm, setOpenConfirm] = useState(false);
-  const [openStatus, setOpenStatus] = useState(false);
-  const [itemToDelete, setItemToDelete] = useState<{
-    id: number;
-    type: 'review';
-  } | null>(null);
-  const [statusMessage, setStatusMessage] = useState('');
+  const [itemToDelete, setItemToDelete] = useState<number | null>(null);
 
   const handleAddReview = () => {
     addReview({ ...newReview, id: reviews.length + 1 });
     setNewReview({ id: 0, description: '', name: '', post: '', imageUrl: '' });
   };
 
-  const handleRemoveReview = (id: number) => {
-    removeReview(id);
-    setStatusMessage('Обучение успешно удалено');
-    setOpenStatus(true);
-  };
-
-  const handleOpenConfirm = (item: { id: number; type: 'review' }) => {
-    setItemToDelete(item);
+  const handleOpenConfirm = (id: number) => {
+    setItemToDelete(id);
     setOpenConfirm(true);
   };
 
@@ -90,18 +43,11 @@ export default function ReviewsAdmin() {
     setItemToDelete(null);
   };
 
-  const handleCloseStatus = () => {
-    setOpenStatus(false);
-    setStatusMessage('');
-  };
-
   const handleConfirmDelete = () => {
-    if (itemToDelete) {
-      if (itemToDelete.type === 'review') {
-        handleRemoveReview(itemToDelete.id);
-      }
+    if (itemToDelete !== null) {
+      removeReview(itemToDelete);
+      setOpenConfirm(false);
     }
-    handleCloseConfirm();
   };
 
   return (
@@ -114,24 +60,28 @@ export default function ReviewsAdmin() {
         <Typography variant="h3" mt={4} mb={4}>
           Добавить новый отзыв:
         </Typography>
-        <StyledTextField
+        <TextField
+          className="adminTextField"
           label="Имя"
           value={newReview.name}
           onChange={(e) => setNewReview({ ...newReview, name: e.target.value })}
         />
-        <StyledTextField
+        <TextField
+          className="adminTextField"
           label="Должность"
           value={newReview.post}
           onChange={(e) => setNewReview({ ...newReview, post: e.target.value })}
         />
-        <StyledTextField
+        <TextField
+          className="adminTextField"
           label="Описание"
           value={newReview.description}
           onChange={(e) =>
             setNewReview({ ...newReview, description: e.target.value })
           }
         />
-        <StyledTextField
+        <TextField
+          className="adminTextField"
           label="URL изображения"
           value={newReview.imageUrl}
           onChange={(e) =>
@@ -181,9 +131,7 @@ export default function ReviewsAdmin() {
               <Typography>{review.description}</Typography>
             </Box>
             <IconButton
-              onClick={() =>
-                handleOpenConfirm({ id: review.id, type: 'review' })
-              }
+              onClick={() => handleOpenConfirm(review.id)}
               color="error"
             >
               <DeleteIcon />
@@ -192,101 +140,39 @@ export default function ReviewsAdmin() {
         ))}
       </Box>
 
-      <Modal
+      <ConfirmModal
         open={openConfirm}
         onClose={handleCloseConfirm}
-        closeAfterTransition
+        title="Подтверждение удаления"
+        description="Вы уверены, что хотите удалить этот отзыв?"
       >
-        <Fade in={openConfirm}>
-          <Box sx={style}>
-            <Typography variant="h6" component="h2">
-              Подтверждение удаления
-            </Typography>
-            <Typography sx={{ mt: 2 }}>
-              Вы уверены, что хотите удалить этот элемент?
-            </Typography>
-            <Stack
-              sx={{
-                flexDirection: 'row',
-                justifyContent: 'center',
-                gap: 2,
-                mt: 4,
-              }}
-            >
-              <Button
-                onClick={handleCloseConfirm}
-                sx={{
-                  backgroundColor: '#090E16',
-                  color: 'white',
-                  borderRadius: '16px',
-                  '&:hover': {
-                    backgroundColor: '#7B4AE2',
-                    '& .MuiLink-root': {
-                      color: '#7B4AE2',
-                      '& .MuiTypography-root': {
-                        color: '#7B4AE2',
-                      },
-                    },
-                  },
-                }}
-              >
-                Отмена
-              </Button>
-              <Button
-                onClick={handleConfirmDelete}
-                sx={{
-                  backgroundColor: '#090E16',
-                  color: 'white',
-                  borderRadius: '16px',
-                  '&:hover': {
-                    backgroundColor: '#7B4AE2',
-                    '& .MuiLink-root': {
-                      color: '#7B4AE2',
-                      '& .MuiTypography-root': {
-                        color: '#7B4AE2',
-                      },
-                    },
-                  },
-                }}
-              >
-                Удалить
-              </Button>
-            </Stack>
-          </Box>
-        </Fade>
-      </Modal>
-
-      <Modal open={openStatus} onClose={handleCloseStatus} closeAfterTransition>
-        <Fade in={openStatus}>
-          <Box sx={style}>
-            <Typography variant="h6" component="h2">
-              Статус удаления
-            </Typography>
-            <Typography sx={{ mt: 2 }}>{statusMessage}</Typography>
-            <Box sx={{ display: 'flex', justifyContent: 'center', mt: 4 }}>
-              <Button
-                onClick={handleCloseStatus}
-                sx={{
-                  backgroundColor: '#090E16',
-                  color: 'white',
-                  borderRadius: '16px',
-                  '&:hover': {
-                    backgroundColor: '#7B4AE2',
-                    '& .MuiLink-root': {
-                      color: '#7B4AE2',
-                      '& .MuiTypography-root': {
-                        color: '#7B4AE2',
-                      },
-                    },
-                  },
-                }}
-              >
-                Закрыть
-              </Button>
-            </Box>
-          </Box>
-        </Fade>
-      </Modal>
+        <Button
+          onClick={handleCloseConfirm}
+          sx={{
+            backgroundColor: '#090E16',
+            color: 'white',
+            borderRadius: '16px',
+            '&:hover': {
+              backgroundColor: '#7B4AE2',
+            },
+          }}
+        >
+          Отмена
+        </Button>
+        <Button
+          onClick={handleConfirmDelete}
+          sx={{
+            backgroundColor: '#090E16',
+            color: 'white',
+            borderRadius: '16px',
+            '&:hover': {
+              backgroundColor: '#7B4AE2',
+            },
+          }}
+        >
+          Удалить
+        </Button>
+      </ConfirmModal>
     </Stack>
   );
 }
