+++ src/components/contexts/AuthContext.tsx
@@ -1,4 +1,4 @@
-import React, { createContext, useContext, useState } from 'react';
+import React, { createContext, useState } from 'react';
 
 interface AuthContextType {
   isAuthenticated: boolean;
@@ -31,10 +31,4 @@ export const AuthProvider: React.FC<{ children: React.ReactNode }> = ({
   );
 };
 
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (!context) {
-    throw new Error('useAuth необходимо использовать в рамках AuthProvider');
-  }
-  return context;
-};
+export default AuthContext;
